AadConfig moved into models
