+++ src/models/AadConfig.ts
