+++ src/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './authConfig';
+export * from '../models/AadConfig';
 export * from './nameToAltId';
 export * from './groups-mapper';
 export * from './challenges-mapper';
